% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_percentiles}
\alias{calculate_percentiles}
\title{Calculating percentiles}
\usage{
calculate_percentiles(data, es, se, weighted, probs)
}
\arguments{
\item{data}{A dataframe.}

\item{es}{A string with the name of the effect sizes column.}

\item{se}{A string with the name of the std errors column.}

\item{weighted}{An argument which defaults to FALSE in the esd_* functions
indicating whether the distribution should be weighted by the inverse
variance.}

\item{probs}{A vector of percentiles to compute. Will be automatically
provided by the esd_* functions based on the pre-specified \code{method}
argument.}
}
\value{
A single row of
}
\description{
Calculating percentiles
}
\examples{
calculate_percentiles(ot_dat, yi, sei, weighted, probs)
}
\keyword{internal}
