% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_percentiles_ci}
\alias{calculate_percentiles_ci}
\title{Calculating percentiles with 95\% CIs based on bootstrapping}
\usage{
calculate_percentiles_ci(data, es, se, probs, weighted, n_bootstrap)
}
\arguments{
\item{data}{A dataframe.}

\item{es}{A string with the name of the effect sizes column.}

\item{se}{A string with the name of the std errors column.}

\item{probs}{A vector of percentiles to compute. Will be automatically
given by the esd_* functions based on the pre-specified \code{method}
argument.}

\item{n_bootstrap}{Number of bootstrapped samples. Will be automatically
given by the esd_* functions, but can be changed by the user.}

\item{weightedAn}{argument which defaults to FALSE in the esd_* functions
indicating whether the distribution should be weighted by the inverse
variance.}
}
\value{
A row containing the original percentiles and the bootstrapped
percentiles (mean, median, and 95\% CI for each percentile)
}
\description{
Calculating percentiles with 95\% CIs based on bootstrapping
}
\examples{
calculate_percentiles_ci(ot_dat, yi, sei, weighted, probs)

}
\keyword{internal}
