% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_to_se.R
\name{ci_to_se}
\alias{ci_to_se}
\title{Converting 95\% CI to SE}
\usage{
ci_to_se(lbound, ubound)
}
\arguments{
\item{lbound}{lower bound of 95\% confidence interval}

\item{ubound}{upper bound of 95\% confidence interval}
}
\value{
standard error
}
\description{
Converting 95\% CI to SE
}
\examples{
ci_to_se(-0.12, 1.71)
}
