% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_perc.R
\name{esd_perc}
\alias{esd_perc}
\title{Title}
\usage{
esd_perc(df, es, value)
}
\arguments{
\item{df}{Dataset}

\item{es}{Column name of effect sizes}

\item{value}{The value of the effect size of interest}
}
\value{
A value which describes to which percentile in the ESD the inputted
value corresponds
}
\description{
Title
}
\examples{
esd_perc(dat, yi, 0.2)
}
