% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_plot.R
\name{esd_plot}
\alias{esd_plot}
\title{Plotting effect size distributions.}
\usage{
esd_plot(
  df,
  es,
  se = NULL,
  es_type = "Effect size",
  grouping_var = NULL,
  min_group_size = 20,
  weighted = FALSE,
  method = NULL,
  abs = FALSE,
  ci = FALSE,
  sesoi = NULL,
  bin_width = 0.1,
  n_bootstrap = 1000
)
}
\arguments{
\item{df}{A dataframe.}

\item{es}{The effect size column.}

\item{se}{The standard error column (defaults to NULL).}

\item{es_type}{A string specifying the type of effect size (for x-axis
label).}

\item{grouping_var}{Column name of grouping variable.}

\item{min_group_size}{Sets the minimum amount of effect sizes in each group
as specified by the \code{grouping_var} variable. Defaults to 20.}

\item{weighted}{Defaults to FALSE. If set to TRUE, will weight all effect
sizes by their standard error (requires \code{se} to be defined).}

\item{method}{Defaults to NULL. Can be set to "quads" (.25, .5, .75) or
"thirds" (.1665, .5, .8335) to plot vertical lines based on the respective
benchmarks.}

\item{ci}{Defaults to FALSE. If set to TRUE, will plot 95\% CIs for each
benchmark (requires \code{method} to be defined).}

\item{sesoi}{Defaults to NULL. When given a value, will calculate the
percentage of (weighted) effect sizes larger (or equal to) this value.}

\item{bin_width}{Width of the histogram bins. Defaults to 0.1.}

\item{n_bootstrap}{Number of bootstrapped samples for benchmark 95\% CIs.}
}
\value{
A 'ggplot' object.
}
\description{
Plotting effect size distributions.
}
\examples{
esd_plot(ot_dat, yi)
esd_plot(ot_dat, yi, grouping_var = group)
}
