% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_plot_pba.R
\name{esd_plot_pba}
\alias{esd_plot_pba}
\title{Title}
\usage{
esd_plot_pba(
  df = NULL,
  es = NULL,
  se = NULL,
  lim_obj = NULL,
  es_type = "Effect size",
  weighted = FALSE,
  sesoi = NULL,
  sum_es = TRUE,
  method = NULL,
  abs = FALSE,
  ci = FALSE,
  sum_es_type = "diamond",
  bin_width = 0.1,
  n_bootstrap = 1000
)
}
\arguments{
\item{df}{A dataframe.}

\item{es}{The effect size column.}

\item{se}{The standard error column (defaults to NULL).}

\item{lim_obj}{A 'limitmeta' object.}

\item{es_type}{A string specifying the type of effect size (for x-axis
label).}

\item{weighted}{Defaults to FALSE. If set to TRUE, will weight all effect
sizes by their standard error (requires \code{se} to be defined).}

\item{sesoi}{Defaults to NULL. When given a value, will calculate the
percentage of (weighted) effect sizes larger (or equal to) this value.}

\item{sum_es}{Defaults to TRUE. Plots the summary effect
size and 95\% CI for both the distribution of raw effect sizes as well as
the distribution of adjusted effect sizes.}

\item{method}{Defaults to NULL. Can be set to "quads" (.25, .5, .75) or
"thirds" (.1665, .5, .8335) to plot vertical lines based on the respective
benchmarks.}

\item{ci}{Defaults to FALSE. If set to TRUE, will plot 95\% CIs for each
benchmark (requires \code{method} to be defined).}

\item{sum_es_type}{Defaults to "diamond". Sets the type of summary effect
size visualisation. "diamond" will create a diamond shape centred around
the summary effect size, with the width of the diamond showing the 95\% CI.
Can also be set to "dot", which shows a dot centred on the summary effect
size, with lines and brackets showing the width of the 95\% CI.}

\item{bin_width}{Width of the histogram bins. Defaults to 0.1.}

\item{n_bootstrap}{Number of bootstrapped samples for benchmark 95\% CIs.}
}
\value{
A 'ggplot' object.
}
\description{
Title
}
\examples{
esd_plot_pba(ot_dat, yi, sei)
}
