% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_plot_perc.R
\name{esd_plot_perc}
\alias{esd_plot_perc}
\title{Plot effect sizes as a function of percentile}
\usage{
esd_plot_perc(df, es, es_type, method = "quads")
}
\arguments{
\item{df}{Dataset}

\item{es}{Column name for effect sizes}

\item{es_type}{String describing the type of effect size used.}

\item{method}{Defaults to 'quads', can also be thirds}
}
\value{
ggplot object
}
\description{
Plot effect sizes as a function of percentile
}
\examples{
esd_plot_perc(df, es, "Cohen's d")
}
