% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_table.R
\name{esd_table}
\alias{esd_table}
\title{Create a table of effect size benchmarks.}
\usage{
esd_table(
  df,
  es,
  se = NULL,
  grouping_var = NULL,
  min_group_size = 3,
  weighted = FALSE,
  method = "quads",
  ci = FALSE,
  n_bootstrap = 1000,
  bowley = FALSE,
  csv_write = FALSE,
  path_file_name = "esd_table.csv",
  ndec = 2
)
}
\arguments{
\item{df}{A dataframe.}

\item{es}{The effect size column.}

\item{se}{The standard error column (defaults to NULL).}

\item{grouping_var}{Column name of grouping variable.}

\item{min_group_size}{Sets the minimum amount of effect sizes in each group
as specified by the \code{grouping_var} variable. Defaults to 20.}

\item{weighted}{Defaults to FALSE. If set to TRUE, will weight all effect
sizes by their standard error (requires \code{se} to be defined).}

\item{method}{Defaults to "quads". Can be set to "quads" (.25, .5, .75) or
"thirds" (.1665, .5, .8335) to plot vertical lines based on the respective
benchmarks.}

\item{ci}{Defaults to FALSE. If set to TRUE, will plot 95\% CIs for each
benchmark (requires \code{method} to be defined).}

\item{n_bootstrap}{Number of bootstrapped samples for benchmark 95\% CIs.}

\item{bowley}{Defaults to FALSE. Will calculate Bowley's skewness when set
to TRUE.}

\item{csv_write}{Defaults to FALSE. Will write the outputted table as a csv.}

\item{path_file_name}{A string containing the directory to which the .csv
file will be saved, including the title of the .csv file (has to end in
'.csv').}

\item{ndec}{Number of decimals. Defaults to 2.}
}
\value{
A table of class 'data.frame'.
}
\description{
Create a table of effect size benchmarks.
}
\examples{
esd_table(ot_dat, yi)
esd_table(ot_dat, yi, grouping_var = group)
}
