% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESDist-package.R
\docType{package}
\name{ESDist-package}
\alias{ESDist}
\alias{ESDist-package}
\title{ESDist: Calculate and plot effect size distributions}
\description{
Calculates and plots the effect size distribution for meta analyses.
}
\author{
\strong{Maintainer}: Bernt Glaser \email{berntgla@uio.no}

}
\keyword{internal}
