% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_perc.R
\name{esd_perc}
\alias{esd_perc}
\title{Calculate the percentile of an ESD corresponding to a specific value}
\usage{
esd_perc(df, es, value, method = "exclusive")
}
\arguments{
\item{df}{Dataset}

\item{es}{Column name of effect sizes}

\item{value}{The value of the effect size of interest}

\item{method}{Specifies method, defaults to 'exclusive' but can also be
'inclusive' or 'grouped'.}
}
\value{
A value which describes to which percentile in the ESD the inputted
value corresponds
}
\description{
Calculate the percentile of an ESD corresponding to a specific value
}
\examples{
esd_perc(ot_dat, yi, 0.2)
}
