% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_plot.R
\name{esd_plot}
\alias{esd_plot}
\title{Creating a plot for an effect size distribution}
\usage{
esd_plot(
  df,
  es,
  es_type,
  pop_es = NULL,
  method = FALSE,
  mean = FALSE,
  bin_width = 0.1
)
}
\arguments{
\item{df}{Dataset}

\item{es}{Column name of effect sizes}

\item{es_type}{A string describing the type of effect size used (e.g.,
"Cohen's d")}

\item{pop_es}{A numeric argument that corresponds to the population ES of
interest. This will split the histogram into two parts around the inputted
value.}

\item{method}{Defaults to 'none', can also be 'thirds' for 16.65th, 50th, and
83.35th percentiles, or 'quads' for 25th, 50th, and 75th percentiles.}

\item{bin_width}{Sets the bin width for the histogram}
}
\value{
A ggplot object
}
\description{
Creating a plot for an effect size distribution
}
\examples{
esd_plot(df, es, es_type = "Cohen's d", method = "thirds", pop)
}
