% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_plot.R
\name{esd_plot}
\alias{esd_plot}
\title{Creating a plot for an effect size distribution}
\usage{
esd_plot(df, es, es_type, method = "quads", bin_width = 0.05)
}
\arguments{
\item{df}{Dataset}

\item{es}{Column name of effect sizes}

\item{es_type}{A string describing the type of effect size used (e.g., "Cohen's d")}

\item{method}{Defaults to 'quads', can also be 'thirds'}

\item{bin_width}{Sets the bin width for the histogram}
}
\value{
A ggplot object
}
\description{
Creating a plot for an effect size distribution
}
\examples{
esd_plot(df, es, es_type = "Cohen's d", method = thirds)
}
