% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_table.R
\name{esd_table}
\alias{esd_table}
\title{Creating a table for an effect size distribution}
\usage{
esd_table(df, es, grouping_var = NULL, method = "quads", csv_write = FALSE)
}
\arguments{
\item{df}{Dataset}

\item{es}{Column name of effect sizes}

\item{grouping_var}{Column name of grouping variable}

\item{method}{Defaults to 'quads', can also be 'thirds'}

\item{csv_write}{Defaults to FALSE. Will write the outputted table as a csv
when set to TRUE.}
}
\value{
a matrix
}
\description{
Creating a table for an effect size distribution
}
\examples{
esd_table(df, es, grouping_var = group, method = "thirds")


}
