% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_table_pba.R
\name{esd_table_pba}
\alias{esd_table_pba}
\title{Adjusting an ESD for publication bias}
\usage{
esd_table_pba(
  df,
  es,
  adj_es,
  grouping_var = NULL,
  method = "quads",
  csv_write = FALSE
)
}
\arguments{
\item{df}{Dataset}

\item{es}{The name of the column containing the effect sizes}

\item{adj_es}{Adjusted effect size. Can be a single number, or a vector of
numbers that correspond to the adjusted effect sizes for each group in
alphabetical order, with the overall adjusted effect size at the end.}

\item{grouping_var}{Column name of grouping variable}

\item{method}{Defaults to 'quads', can also be 'thirds'}

\item{csv_write}{Defaults to FALSE. Will write the outputted table as a csv
when set to TRUE.}
}
\value{
a matrix
}
\description{
Adjusting an ESD for publication bias
}
\examples{
esd_table_pba(df, es, 0.2, grouping_var = group, method = "thirds")

}
